requirements/base.txt
Django==1.6.1
South==0.8.4
ipython
django-extensions==1.2.5
djangorestframework==2.3.10
facebook-sdk==0.4.0
psycopg2==2.5.1
python-mimeparse==0.1.4
six==1.4.1
wsgiref==0.1.2
django-staticfiles-fontawesome==2.0.0
django-endless-pagination==2.0
pyreadline==2.0
pytz==2013.9
requests==2.1.0
unidecode==0.04.14
django-testutils
numpy==1.8.1
pandas==0.13.1
django-pandas==0.1.0
python-dateutil==1.5
django-tastypie==0.9.15
-e git://github.com/rmaceissoft/django-facebook-realtime.git@e6b9f5625084ee1a4f019010e905f5b638f795ea#egg=django-facebook-realtime

-r oknesset.txt